+++ packages/ui/src/legacy/components/ThemeToggle/ThemeToggle.tsx
@@ -1,7 +1,7 @@
 import { MoonIcon, SunIcon } from '@heroicons/react/24/outline';
 import { twMerge } from 'tailwind-merge';
 
-import { useTheme } from '../../hooks/useTheme';
+import { useTheme } from '@/hooks/useTheme';
 
 export type ThemeToggleProps = {
   className?: string;